Restructured pdfcreation module
