+++ kmuhelper/modules/pdfgeneration/__init__.py
@@ -1,0 +1,1 @@
+from kmuhelper.modules.pdfgeneration.order import PDFOrder
