+++ kmuhelper/modules/pdfgeneration/order/__init__.py
@@ -1,0 +1,1 @@
+from kmuhelper.modules.pdfgeneration.order.generator import PDFOrder
