--- kmuhelper/pdf_generators/__init__.py
@@ -1,1 +1,0 @@
-from kmuhelper.pdf_generators.bestellung import PDFOrder
